+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Select 'Acquisition method' filter from controlled vocabulary list. Refs UIPFPOL-33.
 * Update expense class filter to support multi-select. Refs UIOR-870.
+* Filter orders lines by export date. Refs UIPFPOL-38.
 
 ## [3.0.1](https://github.com/folio-org/ui-plugin-find-po-line/tree/v3.0.1) (2021-11-11)
 [Full Changelog](https://github.com/folio-org/ui-plugin-find-po-line/compare/v3.0.0...v3.0.1)